sec: security hardened, userid used only in token form

+++ backend/routes/appointment.routes.js
@@ -2,11 +2,11 @@ const { json } = require("express");
 const Appointment = require("../models/appointment");
 const router = require("express").Router();
 const checkIfAuthenticated = require("../JWTtoken/Validator").checkIfAuthenticated;
+const getLoggedInUserIdFromLogin = require("../JWTtoken/Validator").getLoggedInUserIdFromLogin;
 
-router.route("/getByUser/:id").get(checkIfAuthenticated,(req, res) =>
-{
-    const userid = req.params.id.split("=")[1];
-    Appointment.find({ 'userid': userid }, (error, data)=> {
+router.route("/getByUser").get(checkIfAuthenticated, (req, res) => {
+    const userId = getLoggedInUserIdFromLogin(req.headers.authorization)
+    Appointment.find({ 'userid': userId }, (error, data)=> {
         if (error) {
             return error;
         } else {
@@ -15,9 +15,16 @@ router.route("/getByUser/:id").get(checkIfAuthenticated,(req, res) =>
     })
 });
 
-router.route("/create").post(checkIfAuthenticated,(req, res) =>
-{
-    Appointment.create(req.body, (error, data) => {
+router.route("/create").post(checkIfAuthenticated,(req, res) =>{
+    const userId = getLoggedInUserIdFromLogin(req.headers.authorization)
+    createAppointment = {
+        title: req.body.title,
+        body: req.body.body,
+        date: req.body.date,
+        userid: userId
+    }
+    console.log(createAppointment);
+    Appointment.create(createAppointment, (error, data) => {
         if (error) {
             return next(error)
         } else {
@@ -26,19 +33,27 @@ router.route("/create").post(checkIfAuthenticated,(req, res) =>
     })
 });
 
-router.route("/delete/:id").delete(checkIfAuthenticated,(req, res) => {
+router.route("/delete/:id").delete(checkIfAuthenticated, (req, res) => {
+    const userId = getLoggedInUserIdFromLogin(req.headers.authorization)
     const id = req.params.id.split("=")[1];
-    Appointment.deleteOne({ '_id': id }, (error, data) => {
+    Appointment.deleteOne({ '_id': id, 'userid': userId }, (error, data) => {
         if (error) {
             return error;
         } else {
-            res.sendStatus(200);
+            res.json(data);
         }
     })
 });
 
-router.route("/update").put(checkIfAuthenticated,(req, res) => {
-    Appointment.updateOne({"_id": req.body._id}, req.body, (error, data) => {
+router.route("/update").put(checkIfAuthenticated, (req, res) => {
+    const userId = getLoggedInUserIdFromLogin(req.headers.authorization)
+    createAppointment = {
+        title: req.body.title,
+        body: req.body.body,
+        date: req.body.date,
+        userid: userId
+    }
+    Appointment.updateOne({"_id": req.body._id, 'userid': userId}, createAppointment, (error, data) => {
         console.log(req.body);
         if (error) {
             return error;
@@ -48,9 +63,10 @@ router.route("/update").put(checkIfAuthenticated,(req, res) => {
     })
 });
 
-router.route("/getById/:id").get(checkIfAuthenticated,(req, res) => {
+router.route("/getById/:id").get(checkIfAuthenticated, (req, res) => {
+    const userId = getLoggedInUserIdFromLogin(req.headers.authorization)
     const id = req.params.id.split("=")[1];
-    Appointment.findById({'_id': id}, (error, data) => {
+    Appointment.findById({'_id': id, 'userid': userId}, (error, data) => {
         if (error) {
             return error;
         } else {
